Move hooks to corresponding context directories
