--- src/hooks/index.js
@@ -1,2 +1,0 @@
-export { useHabitsData } from './use-habits-data';
-export { useCheckmarksData } from './use-checkmarks-data';
